Update HUDS.xml
Forgot to cancel the fallback Disciplined Fist timeout when using Anatman. This would cause Disciplined Fist to display 0 when it's duration would have ended without Anatman. This should be the very last update for Monk.

The fallback timeout was needed for near full instances as the non-network logs would cull some combat messages. Not sure if that's still the case in Endwalker.